Refactor code with dao
